// maeri/src/app/models/purchase.model.ts
export interface Purchase {
  AgenceCommande: any[];
  adminId: string;
  articles: any[];
  created: any;
  delivery: boolean;
  desable: boolean;
  hours: string;
  managerConfirm: boolean;
  purchaseDate: string;
  quantity: number;
  scConfirm: boolean;
  senderId: string;
  store: any;
  storeId: string;
  swConfirm: boolean;
  saConfirm: boolean;
  totalPrice: number;
  _id: string;
  wareHouseRefueling: boolean;
}